Implement AbstractMutableSlottedUnionFind and integrate with SlottedUnionFindBuilder
